Move access control logic into User.
