+++ app/controllers/application_controller.rb
@@ -10,7 +10,7 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
   
-#  def current_user
-#    
-#  end
+  def current_user
+    User.new(request)    
+  end
 end
