+++ app/views/catalog/show.erb
@@ -73,7 +73,7 @@
     <a name="player"></a>
     <% if !@pbcore.media_srcs.empty? && 
         !@pbcore.outside_url &&
-        AccessControl.authorized_ip?(request.remote_ip) %>
+        current_user.onsite? # TODO: replace with cancan %>
       <%= render partial: 'catalog/player' %>
     <% else %>
       <img src="<%= @pbcore.img_src %>" alt="thumbnail of <%= @pbcore.title %>">
