+++ lib/geo_i_p_country.rb
@@ -1,6 +1,5 @@
 require 'singleton'
 require 'geoip'
-require_relative 'access_control'
 
 class GeoIPCountry
   include Singleton
@@ -10,7 +9,6 @@ class GeoIPCountry
   end
   
   def country_code(ip)
-    return 'US' if AccessControl.authorized_ip?(ip) # WGBH not in geolocate database.
     @geo_ip.country(ip).country_code2
   end
 end
